cmit_automaton/requirements.txt
selenium
pandas
webdriver-manager
pyyaml